create components for code encapsulation
