+++ src/loaders/Loader.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import logo from './logo.svg';
+
+const Loader = () => (
+  <div className="App">
+    <header className="App-header">
+      <img src={logo} className="App-logo" alt="logo" />
+      <p>Loading...</p>
+    </header>
+  </div>
+);
+
+export default Loader;
